refactor(retention): move insight to its own folder
