+++ products/endpoints/backend/insight_transformers.py
@@ -19,7 +19,7 @@ from posthog.models.team import Team
 
 if TYPE_CHECKING:
     from posthog.hogql_queries.insights.lifecycle.lifecycle_query_runner import LifecycleQueryRunner
-    from posthog.hogql_queries.insights.retention_query_runner import RetentionQueryRunner
+    from posthog.hogql_queries.insights.retention.retention_query_runner import RetentionQueryRunner
     from posthog.hogql_queries.insights.trends.trends_query_runner import TrendsQueryRunner
 
     InsightRunner = Union["LifecycleQueryRunner", "RetentionQueryRunner", "TrendsQueryRunner"]
